refactor and rename JBoxContainer for APIs
