+++ engine/src/juliabox/handlers/admin.py
@@ -5,7 +5,7 @@ import isodate
 from juliabox.cloud import Compute
 from juliabox.jbox_util import JBoxCfg
 from handler_base import JBoxHandler
-from juliabox.jbox_container import JBoxContainer
+from juliabox.interactive import SessContainer
 from juliabox.jbox_tasks import JBoxAsyncJob
 from juliabox.db import JBoxUserV2, JBoxDynConfig, JBPluginDB
 from juliabox.api import APIContainer
@@ -23,7 +23,7 @@ class AdminHandler(JBoxHandler):
         is_admin = sessname in JBoxCfg.get("admin_sessnames", [])
         manage_containers = is_admin or user.has_role(JBoxUserV2.ROLE_MANAGE_CONTAINERS)
         show_report = is_admin or user.has_role(JBoxUserV2.ROLE_ACCESS_STATS)
-        cont = JBoxContainer.get_by_name(sessname)
+        cont = SessContainer.get_by_name(sessname)
 
         if cont is None:
             self.send_error()
@@ -46,6 +46,7 @@ class AdminHandler(JBoxHandler):
         if user.has_resource_profile(JBoxUserV2.RES_PROF_JULIA_PKG_PRECOMP):
             jimg_type = JBoxUserV2.RES_PROF_JULIA_PKG_PRECOMP
 
+        expire = JBoxCfg.get('interactive.expire')
         d = dict(
             manage_containers=manage_containers,
             show_report=show_report,
@@ -53,11 +54,11 @@ class AdminHandler(JBoxHandler):
             user_id=user_id,
             created=isodate.datetime_isoformat(cont.time_created()),
             started=isodate.datetime_isoformat(cont.time_started()),
-            allowed_till=isodate.datetime_isoformat((cont.time_started() + timedelta(seconds=JBoxCfg.get('expire')))),
+            allowed_till=isodate.datetime_isoformat((cont.time_started() + timedelta(seconds=expire))),
             mem=cont.get_memory_allocated(),
             cpu=cont.get_cpu_allocated(),
             disk=cont.get_disk_allocated(),
-            expire=JBoxCfg.get('expire'),
+            expire=expire,
             juliaboxver=juliaboxver,
             jimg_type=jimg_type
         )
@@ -94,7 +95,7 @@ class AdminHandler(JBoxHandler):
     def handle_if_logout(self, cont):
         logout = self.get_argument('logout', False)
         if logout == 'me':
-            JBoxContainer.invalidate_container(cont.get_name())
+            SessContainer.invalidate_container(cont.get_name())
             JBoxAsyncJob.async_backup_and_cleanup(cont.dockid)
             response = {'code': 0, 'data': ''}
             self.write(response)
@@ -185,10 +186,10 @@ class AdminHandler(JBoxHandler):
     def get_upgrade_available(cont):
         cont_images = cont.get_image_names()
         juliaboxver = cont_images[0]
-        if (JBoxContainer.DCKR_IMAGE in cont_images) or ((JBoxContainer.DCKR_IMAGE + ':latest') in cont_images):
+        if (SessContainer.DCKR_IMAGE in cont_images) or ((SessContainer.DCKR_IMAGE + ':latest') in cont_images):
             upgrade_available = None
         else:
-            upgrade_available = JBoxContainer.DCKR_IMAGE
+            upgrade_available = SessContainer.DCKR_IMAGE
             if ':' not in upgrade_available:
                 upgrade_available += ':latest'
         return juliaboxver, upgrade_available
