+++ engine/src/juliabox/interactive/__init__.py
@@ -1,0 +1,2 @@
+__author__ = 'tan'
+from sess_container import SessContainer
